move language-picker to user-settings module
