+++ src/modules/user-settings/components/preferences-tab.tsx
@@ -2,9 +2,9 @@ import { useI18n } from '@wordpress/react-i18n';
 import { useState } from 'react';
 import Button from 'src/components/button';
 import { EditorPicker } from 'src/components/editor-picker';
-import { LanguagePicker } from 'src/components/language-picker';
 import { useEditorData } from 'src/hooks/use-editor-data';
 import { useI18nData } from 'src/hooks/use-i18n-data';
+import { LanguagePicker } from 'src/modules/user-settings/components/language-picker';
 
 export const PreferencesTab = ( { onClose }: { onClose: () => void } ) => {
 	const { __ } = useI18n();
